+++ src/cc/include/swcdb/fs/Broker/Protocol/req/AppendSync.h
@@ -23,7 +23,7 @@ class AppendSync final : public BaseSync, public Base {
           StaticBuffer& buffer, FS::Flags flags)
         : Base(
             Buffers::make(
-              Params::AppendReq(smartfd->fd(), (uint8_t)flags),
+              Params::AppendReq(smartfd->fd(), flags),
               buffer,
               0,
               FUNCTION_APPEND, timeout
