+++ src/cc/lib/swcdb/core/comm/Header.cc
@@ -1,4 +1,4 @@
-/* 
+/*
  * SWC-DB© Copyright since 2019 Alex Kashirin <kashirin.alex@gmail.com>
  * License details at <https://github.com/kashirin-alex/swc-db/#license>
  */
@@ -20,14 +20,14 @@ Header::Header(uint64_t cmd, uint32_t timeout)
 }
 SWC_SHOULD_INLINE
 Header::Header(const Header& init_from_req_header)
-              : version(1), header_len(0), 
+              : version(1), header_len(0),
                 flags(init_from_req_header.flags),
-                id(init_from_req_header.id), 
-                timeout_ms(0), 
+                id(init_from_req_header.id),
+                timeout_ms(0),
                 command(init_from_req_header.command),
                 buffers(0), checksum(0) {
 }
-  
+
 Header::~Header() { }
 
 SWC_SHOULD_INLINE
@@ -51,7 +51,7 @@ void Header::set(uint64_t cmd, uint32_t timeout) {
 }
 
 SWC_SHOULD_INLINE
-size_t Header::encoded_length() { 
+size_t Header::encoded_length() {
   header_len = FIXED_LENGTH;
   buffers = 0;
   if(data.size) {
@@ -75,7 +75,7 @@ void Header::encode(uint8_t** bufp) const {
   Serialization::encode_i32(bufp, timeout_ms);
   Serialization::encode_i16(bufp, command);
   Serialization::encode_i8(bufp,  buffers);
-  
+
   if(data.size) {
     data.encode(bufp);
     if(data_ext.size)
@@ -89,8 +89,8 @@ SWC_SHOULD_INLINE
 void Header::decode_prefix(const uint8_t** bufp, size_t* remainp) {
   if (*remainp < PREFIX_LENGTH)
     SWC_THROWF(Error::COMM_BAD_HEADER,
-              "Header size %d is less than the fixed length %d",
-              (int)*remainp, PREFIX_LENGTH);
+              "Header size %lu is less than the fixed length %d",
+              *remainp, PREFIX_LENGTH);
 
   version = Serialization::decode_i8(bufp, remainp);
   header_len = Serialization::decode_i8(bufp, remainp);
@@ -100,12 +100,12 @@ SWC_SHOULD_INLINE
 void Header::decode(const uint8_t** bufp, size_t* remainp) {
   const uint8_t *base = *bufp;
   *bufp += PREFIX_LENGTH;
-    
+
   flags = Serialization::decode_i8(bufp, remainp);
   id = Serialization::decode_i32(bufp, remainp);
   timeout_ms = Serialization::decode_i32(bufp, remainp);
   command = Serialization::decode_i16(bufp, remainp);
-    
+
   if((buffers = Serialization::decode_i8(bufp, remainp))) {
     data.decode(bufp, remainp);
     if(buffers > 1)
@@ -114,7 +114,7 @@ void Header::decode(const uint8_t** bufp, size_t* remainp) {
 
   checksum = Serialization::decode_i32(bufp, remainp);
   if(!Core::checksum_i32_chk(checksum, base, header_len-4))
-    SWC_THROWF(Error::COMM_HEADER_CHECKSUM_MISMATCH, 
+    SWC_THROWF(Error::COMM_HEADER_CHECKSUM_MISMATCH,
               "header-checksum decoded-len=%lu", *bufp-base);
 }
 
@@ -137,18 +137,18 @@ void Header::initialize_from_request(const Header& header) {
 }
 
 void Header::print(std::ostream& out) const {
-  out << "version="     << (int)version
-      << " header_len=" << (int)header_len
-      << " flags="      << (int)flags
-      << " id="         << (int)id
-      << " timeout_ms=" << (int)timeout_ms
-      << " command="    << (int)command
-      << " buffers="    << (int)buffers;
+  out << "version="     << int(version)
+      << " header_len=" << int(header_len)
+      << " flags="      << int(flags)
+      << " id="         << int(id)
+      << " timeout_ms=" << int(timeout_ms)
+      << " command="    << int(command)
+      << " buffers="    << int(buffers);
   if(buffers) {
     data.print(out);
     data_ext.print(out);
   }
-  out << " checksum="   << (int)checksum;
+  out << " checksum="   << int(checksum);
 }
